+++ src/main/java/com/raoulvdberge/refinedstorage/integration/jei/RSJEIPlugin.java
@@ -1,5 +1,6 @@
 package com.raoulvdberge.refinedstorage.integration.jei;
 
+import com.raoulvdberge.refinedstorage.gui.GuiBase;
 import mezz.jei.api.IJeiRuntime;
 import mezz.jei.api.IModPlugin;
 import mezz.jei.api.IModRegistry;
@@ -21,6 +22,8 @@ public class RSJEIPlugin implements IModPlugin {
 
         registry.addRecipeRegistryPlugin(new RecipeRegistryPluginCover());
         registry.addRecipeRegistryPlugin(new RecipeRegistryPluginHollowCover());
+
+        registry.addGhostIngredientHandler(GuiBase.class, new GhostIngredientHandler());
     }
 
     @Override